Added Encryption Layers Post
Brief introduction to the concept of differnet layers where encryption can be applied.

A discussion of each of the layers, what it involves, and the benefits and detractors.

Close with some brief comments on the tradeoffs.

This post should have an updated copy of the image used in the Database Encryption and Tokenization post from 2009(?)